==> code/01_preprocessing.ipynb <==
### Install Required Libraries

#!pip install pandas numpy scikit-learn lightfm tqdm matplotlib seaborn scipy

### Import Libraries

import pandas as pd
import numpy as np
import os
import re
import csv
import ast
import pickle
from datetime import datetime
from sklearn.feature_extraction.text import CountVectorizer
pd.set_option('display.max_columns', None)

### Helper

def create_folder(path):
    """Creates a path within the working directory
    Input: path (str) -> specify the working directory
    Output -> path will be created if not existed
    """
    try: 
        os.mkdir(path) 
    except OSError as error: 
        print(error)  

def extract_year(movie_year):
    """Using regular expression, we can extract the released years for each movie.
    Input: movie_year (str) -> Movie title with years
    Output: released_year (int) -> Released years from the given movie."""
    
    
    # create a pattern to extract movie year
    # if 4 digits within the parenthesis, then extract it
    pattern = re.findall(r'\(([0-9]{4})\)', movie_year)
    
    if pattern:
        global released_year
        released_year = int(pattern[0])
    else:
        pass
    
    return released_year

def extract_movie(val):
    """Using regular expression, we can extract the released years for each movie.
    Input:  val (str) -> movie title
    Output: final_title (str) -> The transformed movie title."""
    
    title = str(val)
    
    title_only = re.sub(r'\(([0-9]{4})\)', '', title)
    final_title = title_only.strip()
    
    return final_title

def change_timestamp(utc_timestamp):
    """Change the date format
    Input: utc_timestamp (int) -> timestamp before transformed
    Output: new_date (str) -> datetime information of reviews after transform"""
    
    date_format = "%Y-%m-%d %H:%M:%S"
    new_date = datetime.utcfromtimestamp(utc_timestamp).strftime(date_format)
    new_date = str(new_date)
    
    return new_date

- assets

asset_path = '../assets'
data_path = '../data'

create_folder(asset_path)
create_folder(data_path)

### Import data

movie_path = '../data/movies.csv'
rating_path = '../data/ratings.csv'

movie = pd.read_csv(movie_path)
rating = pd.read_csv(rating_path)

### Quick check

movie.info()

rating.info()

- check unique values

print(f"In the movie dataset, there are {movie.movieId.nunique()} unique movies.")

print(f"In the rating dataset, there are {rating.userId.nunique()} unique users.")
print(f"In the rating dataset, {rating.movieId.nunique()} movies were rated by various users.")

### Data Transformation

#### Movie

movie.head()

movie.info()

movie.columns

- extract movie names

movie.loc[:, 'clean_movie_title'] = movie.title.apply(extract_movie)

- extract years

movie.loc[:, 'released_year'] = movie.title.apply(extract_year)

- genre

movie.loc[:, 'genre_counts'] = movie.genres.apply(lambda row: len(row.split('|')))

find the number of gernes for each movie

transform some values since we will work on one hot coding

movie.loc[:, 'new_genres'] = movie.genres.apply(lambda row: 'Missing' if 'no genres listed' in str(row) else row)

movie.loc[:, 'new_genres'] = movie.new_genres.apply(lambda row: row.replace('|', ','))

- grouping movie released years

movie.loc[:, 'groupped_released_year']= movie.released_year.apply(lambda year: str(year)[:3] + '0s')

- one hot coding

target = list(movie.new_genres)

cv = CountVectorizer()
genre_to_cv = cv.fit_transform(target)

genre_df = pd.DataFrame(genre_to_cv.toarray(), columns = cv.get_feature_names_out())

genre_df.loc[:, 'movieId'] = list(movie.movieId)

- combine the movie dataset together

final_movie = pd.merge(movie, genre_df, left_on = ['movieId'], right_on = ['movieId'])

#### Rating

rating.head()

rating.info()

- convert the timestamp

rating.loc[:, 'review_datetime'] = rating.timestamp.apply(change_timestamp)

rating.review_datetime = pd.to_datetime(rating.review_datetime)

convert it to the datetime object

rating.loc[:, 'review_year'] = rating.review_datetime.dt.year
rating.loc[:, 'review_quarter'] = rating.review_datetime.dt.quarter
rating.loc[:, 'review_month'] = rating.review_datetime.dt.month
rating.loc[:, 'review_day'] = rating.review_datetime.dt.day
rating.loc[:, 'review_dayname'] = rating.review_datetime.dt.day_name()

get date information from the datetime object

### Join the dataset

join the datasets together to get user-item information

convert movieId as string to join together

- convert the date type from int to str

final_movie.loc[:, 'movieId'] = final_movie.movieId.apply(lambda ID: str(ID))
rating.loc[:, 'movieId'] = rating.movieId.apply(lambda ID: str(ID))

- select features from both datasets

final_movie.columns

featured_movie = final_movie.copy()

rating.columns

movie_features = ['movieId', 'clean_movie_title', 'released_year', 'genre_counts', 'new_genres'\
                  ,'groupped_released_year', 'action', 'adventure', 'animation', 'children', 'comedy'\
                  ,'crime', 'documentary', 'drama', 'fantasy', 'fi', 'film', 'horror', 'imax'\
                  ,'missing', 'musical', 'mystery', 'noir', 'romance', 'sci', 'thriller', 'war', 'western']

rating_features = ['userId', 'movieId', 'rating', 'review_datetime', 'review_year', 'review_quarter'\
                   ,'review_month', 'review_day', 'review_dayname']

movie_selected = final_movie[movie_features]
rating_selected = rating[rating_features]

### Export the dataset

movie_pkl = '../assets/movie.pkl'
rating_pkl = '../assets/rating.pkl'
featured_movie_pkl = '../assets/featured_movie.pkl'

with open(movie_pkl, 'wb') as m:
    pickle.dump(movie_selected, m)

with open(rating_pkl, 'wb') as r:
    pickle.dump(rating_selected, r)

with open(featured_movie_pkl, 'wb') as r:
    pickle.dump(featured_movie, r)
