==> code/04_flask.ipynb <==
import pandas as pd
import numpy as np
import random
import os
import time

from flask import Flask
from flask import request
from flask import escape
from flask import render_template
from flask import url_for
from flask import redirect

#import <python file> function that will output the movie
from hello import hello

app = Flask(__name__, template_folder = 'templates', static_folder = 'static')

@app.route('/')
def homepage():
    return render_template('index.html')



@app.route('/recs', methods=['GET','POST'])
def recs():
    try:
        if request.method == 'POST':
            input_data = request.form
            input_name = input_data.get('name')
            input_genre = input_data.get('genre')
            if (input_name == '') & (input_genre == ''):
                return render_template('modeling.html')
            elif (input_name == ''):
                return render_template('genre_rec.html', x = hello(input_name) y = input_name, z = input_genre) #x= pyfunction(model, movie, dataset, input_name, input_genre)
            else:
                return render_template('final_rec.html', x = hello(input_name), y = input_name, z = input_genre)
            #####
            return render_template('final_rec.html')
            #return f"Woohoo, {input_name} is the best! Recommendations for {input_genre} movies \
            #will be generated in next version."
            #####
            #add above for page return after a movie is generated
        
        
    except Exception as e:
        print(e)
        
    if request.method == 'GET':
        return redirect(url_for('modeling'))




@app.route('/dashboard', methods=['GET','POST'])
def attempt():
    return render_template('myrange_slider.html')



@app.route('/modeling', methods=['POST','GET'])
def modeling():
    return render_template('modeling.html')
    

if __name__=='__main__':
    app.run(host = 'localhost')
