==> code/05_demo.ipynb <==
### Import Libraries

import numpy as np
import pandas as pd
import re
import os
import pickle
from sklearn.preprocessing import LabelEncoder
from lightfm.data import Dataset
from lightfm import LightFM
from lightfm.evaluation import auc_score, precision_at_k
from scipy.sparse import csr_matrix, coo_matrix
from sklearn.preprocessing import LabelEncoder
from tqdm import tqdm 
import matplotlib.pyplot as plt
import seaborn as sns
%matplotlib inline

### Import Data

movie_path = '../assets/final_movie.pkl'
rating_path = '../assets/final_rating.pkl'
model_path = '../assets/model.pkl'
user_item_path = '../assets/user_movie_interaction.pkl'
dataset_path = '../assets/dataset.pkl'
user_feats_path = '../assets/user_feats.pkl'
csr_path = '../assets/csr.pkl'

movie = pd.read_pickle(movie_path)
new_rating = pd.read_pickle(rating_path)

with open(model_path, 'rb') as m:
    model = pickle.load(m)

with open(user_item_path, 'rb') as u:
    user_item_interaction = pickle.load(u)

with open(dataset_path, 'rb') as d:
    dataset = pickle.load(d)

with open(user_feats_path, 'rb') as u:
    user_feats = pickle.load(u)

with open(csr_path, 'rb') as c:
    user_item_csr = pickle.load(c)

### Helpers

def recommend_top5(model, movie, dataset, user_id=None, new_user_feature=None, k=5):
    
    nmovie=movie.set_index('movieId')
    
    max_user_id = max(new_rating.userId)
    if user_id is None:
        user_id = max_user_id +1
    if user_id > max_user_id:
        for genre in unique_genre:
            if genre not in list(new_user_feature.keys()):
                new_user_feature[genre] = 0
        dataset.fit_partial(users=[user_id],user_features=unique_genre)
        new_user_feature = [user_id,new_user_feature]
        new_user_feature = dataset.build_user_features([new_user_feature],normalize=False)

    user_id_map = dataset.mapping()[0][user_id] 
    scores = model.predict(user_id_map, np.arange(n_items),user_features=new_user_feature)
    rank = np.argsort(-scores)
    selected_movie_id =np.array(list(dataset.mapping()[2].keys()))[rank]
    top_items = nmovie.loc[selected_movie_id]

        
    return top_items['clean_movie_title'][:5]

def create_random_user_info(user_id = None):
    
    """This function will create random user information regarding their genre and interests.
    Input: 
    user_id (None): If None, then we assume this is a new user.
                    If we specify the user information, then it will overwrite their previous information.
    unique_genre (list): contains unique genre in the dataset. 
                           
    Output:
    user_info_dict (dict): It will contain user information regarding the genres and their interests.
    """
    
    # create an empty dictionary to save the results
    user_info_dict = {}
    
    for genre in unique_genre:
        rand_score = round(np.random.rand(1)[0], 4)
        user_info_dict[genre] = rand_score
        
    return user_info_dict

### Test

n_users, n_items = user_item_csr.shape

unique_genre = sorted(set(map(lambda genre: \
                                  genre.strip(),','.join(map(lambda genre: str(genre), movie['new_genres'].values))\
                                  .split(','))))

### Testing

change a bit so that we can see what they've watched and their recommended movies

recommend_top5(model,
            movie,
            dataset,
            user_id = 2)

recommend_top5(model,
            movie,
            dataset,
            user_id = None,
            new_user_feature= {'Comedy':1, 'Mystery': 0.5, 'Adventure':0.8})

recommend_top5(model,
            movie,
            dataset,
            user_id = None,
            new_user_feature= {'Comedy':0.6, 'Mystery': 0.5, 'Adventure':0.8})

- random testing

rand_user_info_1 = create_random_user_info()
rand_user_info_2 = create_random_user_info()
rand_user_info_3 = create_random_user_info()

recommend_top5(model,
            movie,
            dataset,
            user_id = None,
            new_user_feature= rand_user_info_1)

recommend_top5(model,
            movie,
            dataset,
            user_id = None,
            new_user_feature= rand_user_info_2)

recommend_top5(model,
            movie,
            dataset,
            user_id = None,
            new_user_feature= rand_user_info_3)
